fix : 음료추가, 자판기 및 더미재고 생성 로직 분리
